indexing exports to make it easier to keep track of components
