+++ src/components/ImageUploader/index.js
@@ -1,0 +1,1 @@
+export { default as ImageUploader } from "./ImageUploader";
