+++ src/components/SpellUpdater/index.js
@@ -1,0 +1,1 @@
+export { default as SpellUpdater } from "./SpellUpdater";
