+++ src/components/SpellViewer/index.js
@@ -1,0 +1,1 @@
+export { default as SpellViewer } from "./SpellViewer";
